uint64_t, fixing bitwise op with ULL, added tests for encode/decode

+++ 01_software/sim/cpu/decoder.c
@@ -1,10 +1,10 @@
 #include "decoder.h"
 
-unsigned int decode_amount(bus64 a) {
-  unsigned int r = 0;
+uint64_t decode_amount(bus64 a) {
+  uint64_t r = 0;
   for (int i = 0; i < BUS64_WIDTH; ++i) {
     if (bus64_getbit(a, i)) {
-      r |= (1U << i);
+      r |= (1ULL << i);
     }
   } 
   return r;
